Spotify intergration done - refresh token!
Spotify is now running as a nuxt plugin in the client. this means that the player now can be controlled everywhere in the app making it seknifigantly esiser to mangae when in use. this is only the playback api so the normal api still need to be setup for showing artis pages and so on