Player Spotify Intergration done + prep work for other service
Spotify token Refreshing is still broken,

The player is now a class. this makes it 5 times easier to expand support for more players

Also have the first class to the spotify web api in place so work can start on getting the users playlists and such.

also made the background change with the music because why not